requirements.txt
# Core Web Framework
fastapi==0.129.0
uvicorn==0.40.0

# Basic Image Processing (no system dependencies)
pillow>=9.0.0
numpy>=1.21.0

# Data Processing
pandas>=1.3.0

# Database
sqlalchemy>=1.4.0

# Web Scraping
requests>=2.25.0
beautifulsoup4>=4.9.0

# Utilities
python-dotenv>=0.19.0